Resolve issues.
Use instanceOf to resolve class cast exception and wait for layout to be
constructed in PhotoEditActivity.

+++ app/src/main/java/com/zulip/android/util/PhotoHelper.java
@@ -24,6 +24,10 @@ public class PhotoHelper {
      * @param bitmap to be saved as file
      */
     public static String saveBitmapAsFile(String photoPath, Bitmap bitmap) {
+        if (bitmap == null) {
+            return photoPath;
+        }
+
         // delete old bitmap
         File file = new File(photoPath);
         file.delete();
